refactor(restraint): split restraint.rs into restraint/ module + docstrings
restraint.rs was 952 lines (over the 800 cap). Split by shape family into
restraint/{inside,outside,plane,cylinder,gaussian}.rs (each <300 lines),
with the Restraint trait + Box impl + two-scale contract in restraint/mod.rs.
Add a 'write your own restraint' doctest and penalty-formula docstrings to
the opaque kinds (cylinder, gaussian). Pure code move — examples_batch parity
green; public re-exports unchanged.
